Made openshift API user-centric
Store the openshift API facade inside of the user, meaning all of the
operations will be performed through a particular user. I think this
makes sense and should make things easier for the chaos engine to pick a
user and dork with its resources (while that logic is still expressed in
terms of the API facade, just not having to annoyingly pass the user in
each time).